🧪 4XZ5WY release: isolate rebase race test
