+++ .agentplane/tasks/202607131808-4XZ5WY/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607131808-4XZ5WY/prepare-and-publish-patch-release-v0-6-23",
   "created_at": "2026-07-13T18:11:52.210Z",
-  "diffstat_sha256": "sha256:0deac283c760b951f51d830f20c0bbdaa91258c48e008d60491b39fbf74b11c2",
+  "diffstat_sha256": "sha256:0a5dec8bab8a1049ba55c9c7f360b620c266bcd66b772b7725934e40fd78fa41",
   "last_verified_at": null,
   "schema_version": 1,
   "task_id": "202607131808-4XZ5WY",
